requirements.txt
flask
opencv-python
mediapipe
numpy
pillow
scikit-learn